add conf target for build/conf/*.conf file depends
If a change is made to any of the meta-gateway-ww/conf/*.conf.sample
templates, the user needs to regenerate the poky/build/*.conf file
from the template to pull in the changes.

This Makefile target should allow the user to simply run 'make'
again (or make conf for a more specific target), and have the
conf file automatically regnerated.